refactor(web): extract usage report mutations into svc.ts
